backend/requirements.txt
asgiref==3.2.10
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
Django==2.2.16
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.1
gunicorn==20.0.4
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==21.3
Pillow==9.3.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.11.0
pycparser==2.21
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2020.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.3.1
toml==0.10.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
zipp==3.10.0
